requirements.txt
bs4
lxml
PyPDF2
python-frontmatter
